+++ R/module-auth.R
@@ -9,28 +9,28 @@
 #' @param tags_bottom A \code{tags (div, img, ...)} to be displayed on bottom of the authentication module.
 #' @param background A optionnal \code{css} for authentication background. See example.
 #' @param choose_language \code{logical/character}. Add language selection on top ? TRUE for all supported languages
-#' or a vector of possibilities like \code{c("en", "fr", "pt-BR", "es", "de", "pl")}. If enabled, \code{input$shinymanager_language} is created
+#' or a vector of possibilities like \code{c("en", "fr", "pt-BR", "es", "de", "pl", "ja", "el", "id")}. If enabled, \code{input$shinymanager_language} is created
 #' @param ... : Used for old version compatibility.
-#' 
-#' 
+#'
+#'
 #' @export
 #'
 #' @name module-authentication
 #'
 #' @importFrom htmltools tagList tags singleton
-#' @importFrom shiny NS fluidRow column textInput passwordInput actionButton uiOutput 
+#' @importFrom shiny NS fluidRow column textInput passwordInput actionButton uiOutput
 #'
 #' @example examples/module-auth.R
-auth_ui <- function(id, status = "primary", tags_top = NULL, 
-                    tags_bottom = NULL, background = NULL, 
+auth_ui <- function(id, status = "primary", tags_top = NULL,
+                    tags_bottom = NULL, background = NULL,
                     choose_language = NULL, lan = NULL, ...) {
-  
+
   ns <- NS(id)
-  
+
   if(is.null(lan)){
     lan <- use_language()
   }
-  
+
   # patch / message changing tag_img & tag_div
   deprecated <- list(...)
   if("tag_img" %in% names(deprecated)){
@@ -61,7 +61,7 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
             tags$div(
               class = "panel-body",
               {
-                
+
                 choices = lan$get_language()
                 lan_registered <- lan$get_language_registered()
                 if(is.logical(choose_language) && choose_language){
@@ -69,7 +69,7 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
                 } else if(is.character(choose_language)){
                   choices = unique(c(intersect(choose_language, unname(lan$get_language_registered())), lan$get_language()))
                 }
-                
+
                 names(choices) <- choices
                 for(i in 1:length(choices)){
                   ind <- which(lan_registered %in% choices[i])
@@ -77,7 +77,7 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
                     names(choices)[i] <- names(lan_registered)[ind]
                   }
                 }
-                selected = ifelse(lan$get_language() %in% choices, 
+                selected = ifelse(lan$get_language() %in% choices,
                                   lan$get_language(),
                                   choices[1])
                 if(length(choices) == 1){
@@ -153,10 +153,10 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
 #'   \item \strong{expired} : logical, is user has expired ? Always \code{FALSE} if \code{db} doesn't have a \code{expire} column. Optional.
 #'   \item \strong{authorized} : logical, is user can access to his app ? Always \code{TRUE} if \code{db} doesn't have a \code{applications} column. Optional.
 #'  }
-#'  
+#'
 #' @param use_token Add a token in the URL to check authentication. Should not be used directly.
 #' @param lan A language object. See  \code{\link{use_language}}
-#' 
+#'
 #' @export
 #'
 #' @rdname module-authentication
@@ -171,16 +171,16 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
 #' @importFrom htmltools tags
 #' @importFrom shiny reactiveValues observeEvent removeUI updateQueryString insertUI is.reactive icon updateActionButton updateTextInput renderUI
 #' @importFrom stats setNames
-auth_server <- function(input, output, session, 
-                        check_credentials, 
+auth_server <- function(input, output, session,
+                        check_credentials,
                         use_token = FALSE, lan = NULL) {
-  
+
   ns <- session$ns
   jns <- function(x) {
     paste0("#", ns(x))
   }
-  
-  
+
+
   if(!is.reactive(lan)){
     if(is.null(lan)){
       lan <- reactiveVal(use_language())
@@ -188,56 +188,56 @@ auth_server <- function(input, output, session,
       lan <- reactiveVal(lan)
     }
   }
-  
-  
+
+
   observe({
     session$sendCustomMessage(
       type = "focus_input",
       message = list(inputId = ns("user_id"))
     )
   })
-  
+
   observe({
     if(!is.null(input$language)){
-      lan()$set_language(input$language) 
+      lan()$set_language(input$language)
       updateTextInput(session, inputId = "user_id", label = lan()$get("Username:"))
       updateTextInput(session, inputId = "user_pwd", label = lan()$get("Password:"))
       updateActionButton(session, inputId = "go_auth", label = lan()$get("Login"))
-      
+
       session$sendCustomMessage(
         type = "update_auth_title",
         message = list(
-          inputId = ns("shinymanager-auth-head"), 
+          inputId = ns("shinymanager-auth-head"),
           title = lan()$get("Please authenticate")
         )
       )
-      
+
       output$update_shinymanager_language <- renderUI({
         shinymanager_language(lan()$get_language())
       })
-      
+
       output$label_language <- renderUI({
-        tags$p(paste0(lan()$get("Language"), " :"), 
+        tags$p(paste0(lan()$get("Language"), " :"),
                style = "text-align: right; font-style: italic; margin-top:5px")
       })
-      
+
     }
   })
-  
-  
+
+
   authentication <- reactiveValues(result = FALSE, user = NULL, user_info = NULL)
-  
+
   observeEvent(input$go_auth, {
     removeUI(selector = jns("msg_auth"))
     res_auth <- check_credentials(input$user_id, input$user_pwd)
-    
+
     # locked account ?
     locked <- FALSE
     pwd_failure_limit <- as.numeric(get_pwd_failure_limit())
     if(length(pwd_failure_limit) > 0 && !is.na(pwd_failure_limit) && !is.infinite(pwd_failure_limit)){
       locked <- check_locked_account(input$user_id, pwd_failure_limit)
     }
-    
+
     if (isTRUE(res_auth$result) & !locked) {
       removeUI(selector = jns("auth-mod"))
       authentication$result <- TRUE
@@ -245,26 +245,26 @@ auth_server <- function(input, output, session,
       authentication$user_info <- res_auth$user_info
       # token <- generate_token(input$user_id)
       token <- .tok$generate(input$user_id)
-      
+
       if (isTRUE(use_token)) {
         # add_token(token, as.list(res_auth$user_info))
         .tok$add(token, as.list(res_auth$user_info))
         addAuthToQuery(session, token, lan()$get_language())
         session$reload()
       }
-      
+
     } else if (isTRUE(res_auth$result) & locked) {
-      
+
       save_logs_failed(input$user_id, status = "Locked Account")
-      
+
       insertUI(
         selector = jns("result_auth"),
         ui = tags$div(
           id = ns("msg_auth"), class = "alert alert-danger",
-          icon("exclamation-triangle"), lan()$get("Your account is locked")
+          icon("triangle-exclamation"), lan()$get("Your account is locked")
         )
       )
-      
+
     } else {
       if (is.null(res_auth$user_info)) {
         save_logs_failed(input$user_id, status = "Unknown user")
@@ -272,7 +272,7 @@ auth_server <- function(input, output, session,
           selector = jns("result_auth"),
           ui = tags$div(
             id = ns("msg_auth"), class = "alert alert-danger",
-            icon("exclamation-triangle"), lan()$get("Username or password are incorrect")
+            icon("triangle-exclamation"), lan()$get("Username or password are incorrect")
           )
         )
       } else if (isTRUE(res_auth$expired)) {
@@ -281,7 +281,7 @@ auth_server <- function(input, output, session,
           selector = jns("result_auth"),
           ui = tags$div(
             id = ns("msg_auth"), class = "alert alert-danger",
-            icon("exclamation-triangle"), lan()$get("Your account has expired")
+            icon("triangle-exclamation"), lan()$get("Your account has expired")
           )
         )
       } else {
@@ -291,19 +291,19 @@ auth_server <- function(input, output, session,
             selector = jns("result_auth"),
             ui = tags$div(
               id = ns("msg_auth"), class = "alert alert-danger",
-              icon("exclamation-triangle"), lan()$get("You are not authorized for this application")
+              icon("triangle-exclamation"), lan()$get("You are not authorized for this application")
             )
           )
         } else {
-          
+
           save_logs_failed(input$user_id, status = "Wrong pwd")
-          
+
           if(!locked){
             insertUI(
               selector = jns("result_auth"),
               ui = tags$div(
                 id = ns("msg_auth"), class = "alert alert-danger",
-                icon("exclamation-triangle"), lan()$get("Username or password are incorrect")
+                icon("triangle-exclamation"), lan()$get("Username or password are incorrect")
               )
             )
           } else {
@@ -311,7 +311,7 @@ auth_server <- function(input, output, session,
               selector = jns("result_auth"),
               ui = tags$div(
                 id = ns("msg_auth"), class = "alert alert-danger",
-                icon("exclamation-triangle"), lan()$get("Your account is locked")
+                icon("triangle-exclamation"), lan()$get("Your account is locked")
               )
             )
           }
@@ -319,7 +319,7 @@ auth_server <- function(input, output, session,
       }
     }
   }, ignoreInit = TRUE)
-  
+
   return(authentication)
 }