+++ dev/shiny-classic-options/ui.R
@@ -22,6 +22,6 @@ secure_app(
   ),
   tag_img = tags$img(src = "https://www.r-project.org/logo/Rlogo.png", width = 100),
   enable_admin = TRUE,
-  language = "fr"
+  language = "fr", choose_language = T
 )